+++ documentation/changelog.md
@@ -11,6 +11,7 @@ modified.
 ## Kotlin classes 
 * /app/src/main/java/com/epse/gallery/DisplayImage.kt
 * /app/src/main/java/com/epse/gallery/ImagesGrid.kt
+* /app/src/main/java/com/epse/gallery/ImagesFetcher.kt
 * /app/src/main/java/com/epse/gallery/MainActivity.kt
 
 ## Other